object-storage-based/requirements.txt
flask
minio

tensorflow
tensorflow-hub
tensorflow-cpu
pillow
